+++ modules/integration/opensearch-tunnel/app.py
@@ -3,20 +3,18 @@
 
 import os
 
-import aws_cdk
-from aws_cdk import App, CfnOutput
+from aws_cdk import App, CfnOutput, Environment
 
 from stack import TunnelStack
 
-account = os.environ["CDK_DEFAULT_ACCOUNT"]
-region = os.environ["CDK_DEFAULT_REGION"]
-
-deployment_name = os.getenv("ADDF_DEPLOYMENT_NAME", "")
-module_name = os.getenv("ADDF_MODULE_NAME", "")
+# Project specific
+project_name = os.getenv("SEEDFARMER_PROJECT_NAME", "")
+deployment_name = os.getenv("SEEDFARMER_DEPLOYMENT_NAME", "")
+module_name = os.getenv("SEEDFARMER_MODULE_NAME", "")
 
 
 def _param(name: str) -> str:
-    return f"ADDF_PARAMETER_{name}"
+    return f"SEEDFARMER_PARAMETER_{name}"
 
 
 vpc_id = os.getenv(_param("VPC_ID"))
@@ -40,11 +38,11 @@ install_script = os.path.join(project_dir, "install_nginx.sh")
 
 
 def generate_description() -> str:
-    soln_id = os.getenv("ADDF_PARAMETER_SOLUTION_ID", None)
-    soln_name = os.getenv("ADDF_PARAMETER_SOLUTION_NAME", None)
-    soln_version = os.getenv("ADDF_PARAMETER_SOLUTION_VERSION", None)
+    soln_id = os.getenv(_param("SOLUTION_ID"), None)
+    soln_name = os.getenv(_param("SOLUTION_NAME"), None)
+    soln_version = os.getenv(_param("SOLUTION_VERSION"), None)
 
-    desc = "ADDF - Opensearch Tunnel"
+    desc = "Opensearch Tunnel"
     if soln_id and soln_name and soln_version:
         desc = f"({soln_id}) {soln_name}. Version {soln_version}"
     elif soln_id and soln_name:
@@ -56,11 +54,12 @@ app = App()
 
 stack = TunnelStack(
     scope=app,
-    id=f"addf-{deployment_name}-{module_name}",
-    env=aws_cdk.Environment(
-        account=account,
-        region=region,
+    id=f"{project_name}-{deployment_name}-{module_name}",
+    env=Environment(
+        account=os.environ["CDK_DEFAULT_ACCOUNT"],
+        region=os.environ["CDK_DEFAULT_REGION"],
     ),
+    project_name=project_name,
     deployment=deployment_name,
     module=module_name,
     vpc_id=vpc_id,